refactors all actions and helpers to new dir in src
